+++ Examples/Positioning/Sources/Config.swift
@@ -6,13 +6,24 @@
 //  Copyright © 2023 Wemap SAS. All rights reserved.
 //
 
+import WemapCoreSDK
 import WemapPositioningSDK
 
 enum PreferencesKey: String {
-    case positioningVersion
+    case positioningVersion,
+         // Core constants
+         userLocationProjectionOnItineraryEnabled,
+         userLocationProjectionOnGraphEnabled
 }
 
 func customKeysAndValues() -> [String: Any] {
+    
+    // Core
+    CoreConstants.userLocationProjectionOnItineraryEnabled = UserDefaults
+        .bool(forKey: .userLocationProjectionOnItineraryEnabled, defaultValue: CoreConstants.userLocationProjectionOnItineraryEnabled)
+    
+    CoreConstants.userLocationProjectionOnGraphEnabled = UserDefaults
+        .bool(forKey: .userLocationProjectionOnGraphEnabled, defaultValue: CoreConstants.userLocationProjectionOnGraphEnabled)
     
     let specificKeysAndValues: [PreferencesKey: Any] = [
         .positioningVersion: Bundle.positioning.version